refactor(evm): clean up assembler (#858)
This cleans up EVM assembly by reusing assembler state across codegen,
splitting raw byte emission into a configured bytecode assembler, and
moving opcode constants into the `op` module. The byte assembler now
accounts for hardfork-gated PUSH0 availability and keeps alternate
compact push encodings scoped to size optimization, while gas
optimization emits literal PUSH sequences.

The assembler peephole path now mutates in place with stack-style
access, removes unused mark labels, and keeps push offset estimation
aligned with final byte emission. The codegen bytecode fixtures are
updated for the gas-mode literal PUSH behavior.

+++ crates/codegen/src/mir/module.rs
@@ -146,7 +146,7 @@ pub struct StorageSlot {
 #[derive(Clone, Debug)]
 pub struct ImmutableSlot {
     /// Byte offset from the start of the immutable scratch area.
-    pub offset: u64,
+    pub offset: u32,
     /// The type of the value stored.
     pub ty: MirType,
     /// The variable name (for debugging).
